api/pyproject.toml
[project]
name = "pont-facturx"
version = "0.1.0"
description = "PDF to Factur-X conversion service - V1 skeleton"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "celery>=5.4",
  "redis>=5.0",
  "python-multipart>=0.0.9",
  "jinja2>=3.1",
  "lxml>=5.2",
  "pypdf>=5.0",
  "pdfminer.six>=20240706",
  "factur-x>=3.0.0",
  "bcrypt==4.1.3",
  "passlib[bcrypt]",
  "python-jose[cryptography]",
  "google-auth>=2.0.0",
  "email-validator>=2.0.0",
  "requests>=2.31.0",
  "cachetools>=5.0.0",
  "argon2-cffi==23.1.0",
  "passlib==1.7.4",


]

[tool.setuptools]
packages = ["app"]
